Patient 0 & 1
True positive (TP)
The model correctly predicts a positive result. For example, a medical test accurately reports that a patient is sick.
True negative (TN)
The model correctly predicts a negative result. For example, a medical test accurately reports that a patient is not sick.
False positive (FP)
The model incorrectly predicts a positive result. For example, a medical test inaccurately reports that a patient is sick when they are not.
False negative (FN)
The model incorrectly predicts a negative result. For example, a medical test inaccurately reports that a patient is not sick when they are.